Fixed error caused by change to the parent class of Icarus_g2005.
The error I got was that self.compile_args is now a dictionary,
rather than an array. Although self.compile_args_flattened seemed
to work, I wasn't sure if there were other changes in the parent
class that I didn't find.

So this commit was implemented to reduce the reliance on the internal
structure of the parent class.